// packages/agent-toolkit/src/monday-graphql/generated/graphql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A string or number value for comparison */
  CompareValue: { input: any; output: any; }
  /** A date. */
  Date: { input: any; output: any; }
  /** A file */
  File: { input: any; output: any; }
  /** An ISO 8601-encoded datetime (e.g., 2024-04-09T13:45:30Z) */
  ISO8601DateTime: { input: any; output: any; }
  /** A JSON formatted string. */
  JSON: { input: any; output: any; }
  policy__Policy: { input: any; output: any; }
};
// ...
export type BoardView = {
  __typename?: 'BoardView';
  /** The view's access level */
  access_level: BoardViewAccessLevel;
  /** The view's filter */
  filter?: Maybe<Scalars['JSON']['output']>;
  /** The view's filter team id */
  filter_team_id?: Maybe<Scalars['Int']['output']>;
  /** The view's filter user id */
  filter_user_id?: Maybe<Scalars['Int']['output']>;
  /** The view's unique identifier. */
  id: Scalars['ID']['output'];
  /** The view's name. */
  name: Scalars['String']['output'];
  /** The view's settings, the structure varies by view type */
  settings?: Maybe<Scalars['JSON']['output']>;
  /** The view's settings in a string form. */
  settings_str: Scalars['String']['output'];
  /** The view's sort */
  sort?: Maybe<Scalars['JSON']['output']>;
  /** The view's template id if it was duplicated from other view */
  source_view_id?: Maybe<Scalars['ID']['output']>;
  /** The view's tags */
  tags?: Maybe<Array<Scalars['String']['output']>>;
  /** The view's type. */
  type?: Maybe<Scalars['String']['output']>;
  /** Specific board view data (supported only for forms) */
  view_specific_data_str: Scalars['String']['output'];
};

/** The board view access level of the user */
export enum BoardViewAccessLevel {
  /** Edit */
  Edit = 'edit',
  /** View */
  View = 'view'
}
// ...
export type ColumnsConfigInput = {
  /** Order of columns */
  column_order?: InputMaybe<Array<Scalars['String']['input']>>;
  /** Configuration for main board columns */
  column_properties?: InputMaybe<Array<ColumnPropertyInput>>;
  /** Number of floating columns to display */
  floating_columns_count?: InputMaybe<Scalars['Int']['input']>;
  /** Configuration for subitems columns */
  subitems_column_properties?: InputMaybe<Array<ColumnPropertyInput>>;
};
// ...
export type GroupBySettingsInput = {
  /** List of conditions for grouping items */
  conditions: Array<GroupByConditionInput>;
  /** Whether to hide groups with no items */
  hideEmptyGroups?: InputMaybe<Scalars['Boolean']['input']>;
};
// ...
export enum ItemsOrderByDirection {
  /** Ascending order */
  Asc = 'asc',
  /** Descending order */
  Desc = 'desc'
}
// ...
export type ItemsQueryGroup = {
  /** A list of rule groups */
  groups?: InputMaybe<Array<ItemsQueryGroup>>;
  /** The operator to use for the query rules or rule groups. Default: AND */
  operator?: InputMaybe<ItemsQueryOperator>;
  /** A list of rules */
  rules?: InputMaybe<Array<ItemsQueryRule>>;
};

/** Logical operator */
export enum ItemsQueryOperator {
  /** Logical AND */
  And = 'and',
  /** Logical OR */
  Or = 'or'
}

/** Sort the results by specified columns */
export type ItemsQueryOrderBy = {
  column_id: Scalars['String']['input'];
  /** Sort direction (defaults to ASC) */
  direction?: InputMaybe<ItemsOrderByDirection>;
};

/** A rule to filter items by a specific column */
export type ItemsQueryRule = {
  column_id: Scalars['ID']['input'];
  compare_attribute?: InputMaybe<Scalars['String']['input']>;
  compare_value: Scalars['CompareValue']['input'];
  operator?: InputMaybe<ItemsQueryRuleOperator>;
};

/** Rule operator */
export enum ItemsQueryRuleOperator {
  /** Any of the values */
  AnyOf = 'any_of',
  /** Between the two values */
  Between = 'between',
  /** Contains all the terms */
  ContainsTerms = 'contains_terms',
  /** Contains the text */
  ContainsText = 'contains_text',
  /** Ends with the value */
  EndsWith = 'ends_with',
  /** Greater than the value */
  GreaterThan = 'greater_than',
  /** Greater than or equal to the value */
  GreaterThanOrEquals = 'greater_than_or_equals',
  /** Empty value */
  IsEmpty = 'is_empty',
  /** Not empty value */
  IsNotEmpty = 'is_not_empty',
  /** Lower than the value */
  LowerThan = 'lower_than',
  /** Lower than or equal to the value */
  LowerThanOrEqual = 'lower_than_or_equal',
  /** None of the values */
  NotAnyOf = 'not_any_of',
  /** Does not contain the text */
  NotContainsText = 'not_contains_text',
  /** Starts with the value */
  StartsWith = 'starts_with',
  /** Within the last */
  WithinTheLast = 'within_the_last',
  /** Within the next */
  WithinTheNext = 'within_the_next'
}
// ...
export enum SortDirection {
  /** Ascending order */
  Asc = 'ASC',
  /** Descending order */
  Desc = 'DESC'
}
// ...
export type TableViewSettingsInput = {
  /** Column visibility configuration for the board view */
  columns?: InputMaybe<ColumnsConfigInput>;
  /** The group by to apply to the board view */
  group_by?: InputMaybe<GroupBySettingsInput>;
};
// ...
export type CreateViewTableMutationVariables = Exact<{
  boardId: Scalars['ID']['input'];
  name?: InputMaybe<Scalars['String']['input']>;
  filter?: InputMaybe<ItemsQueryGroup>;
  sort?: InputMaybe<Array<ItemsQueryOrderBy> | ItemsQueryOrderBy>;
  tags?: InputMaybe<Array<Scalars['String']['input']> | Scalars['String']['input']>;
  settings?: InputMaybe<TableViewSettingsInput>;
}>;


export type CreateViewTableMutation = { __typename?: 'Mutation', create_view_table?: { __typename?: 'BoardView', id: string, name: string, type?: string | null } | null };